devide raw method and put them in logical functions
